feat: 完善附件管理功能

+++ keen-attachment/src/main/java/com/simple/keen/attachment/model/entity/AttachmentStorage.java
@@ -37,7 +37,7 @@ public class AttachmentStorage {
      * 文件名
      */
     @TableField(value = "storage_size")
-    private Integer storageSize;
+    private long storageSize;
 
     /**
      * 附件标识
